+++ meshmore_sns_app/packages/meshcore/lib/meshcore.dart
@@ -5,4 +5,9 @@
 library;
 
 export 'src/codec/constants.dart';
+export 'src/codec/decode_error.dart';
+export 'src/codec/frame_codec.dart';
+export 'src/codec/inbound.dart';
+export 'src/model/contact.dart';
+export 'src/model/self_info.dart';
 export 'src/transport/transport.dart';